requirements.txt
gradio==4.19.2
pandas
fpdf==1.7.2
matplotlib==3.8.4
numpy==1.26.4
folium==0.16.0
Pillow==10.2.0

# commande a faire cest pip install -r requirements.txt
